[AIP-21] Move hooks.slack_hook providers.slack.hooks.slack
