+++ docs/operators-and-hooks-ref.rst
@@ -961,7 +961,7 @@ These integrations allow you to perform various operations within various servic
 
    * - `Slack <https://slack.com/>`__
      -
-     - :mod:`airflow.hooks.slack_hook`,
+     - :mod:`airflow.providers.slack.hooks.slack`,
        :mod:`airflow.providers.slack.hooks.slack_webhook`
      - :mod:`airflow.operators.slack_operator`,
        :mod:`airflow.contrib.operators.slack_webhook_operator`
